Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -19,7 +19,7 @@ int main() {
     }
     else if(x=='/'){
         result1=a/b;
-        printf("%.4f",result1);
+        printf("%d",(int)result1);
     }
     else{
         printf("error");
